Adjust kafka's maximum poll period. (#2450)
Kafka allows setting the maximum period between two polls. As we allow for 5 minutes of action execution, in a fully loaded invoker it might take 5 minutes for the consumer to poll again. To prevent rebalances this value is adjusted accordingly.

Bumping Kafka's version is needed because that value is exposed only in 0.10.1+.