Fix order of messages being sent by the ContainerPool. (#2421)
The ContainerPool indicates to the ActivationFeed that it can perform more work and as a result gets more messages from the ActivationFeed. In order to properly back-pressure and protect the ContainerPool, it only ever sends such a message iff it changed its internal state to an extent that it is guaranteed to be able to handle more work.

Also add log message to complain about improper buffer.