Add init and wait time to activation record as annotation. (#3053)
Annotate an activation record with two new properties: initTime and waitTime. The former is present for cold starts and reflects the time spent in /init. The latter is the time from the controller receiving the request to the time the invoker initiated a /init or /run for the request against a container. Is it a rough indicator of the queueing time for the request. It is only present for non-sequence activations.