Support CLI built for multi-arch (#3145)
It was also necessary to fiddle with the CLI download capabilities, now that the CLI is being automatically compiled for s390x. The result is not quite as dynamic as one might wish, yet, but it appears to get the job done.

This change is temporary -- responsibility for determining CLI build architectures is being moved toward the CLI build process in a separate repository.