use PureConfig for misc. runtimesManifest configuration (#3259)
Refactor runtimes.json so that the misc. simple configuration
fields are specified using PureConfig (whisk.runtimes),
leaving only the more complex arrays of runtimes and blackboxes
in runtime.json.  Since we expect deployments to frequently
override the set of runtimes, it makes sense to leave them as
being defined by a json file that is not included in the
Invoker/Controller docker image (simplify override).  The flags
moved to PureConfig are simple fields, and thus easier to override
via envvars if necessary.

Restores the ability to easily override bypassPullForLocalImages.

Fixes #3245.