Measure kafka queue in consumer with better exactitude. (#3423)
Measure kafka queue metric directly in the consumer by comparing the offsets instead of relying on the built-in consumer metrics. This method provides more precision compared to the old approach, it was observed that the built-in metrics show certain lagging in case of burst.
Additionally make kamon flush cadency configurable in the application.conf.