Add an optional display name to the invokers. (#3855)
When running the invokers on a container orchestrator, their name in that orchestrator will differ from the topic names used to schedule to them. While the former are usually auto-generated, the latter follow a strict form of a strictly increasing number, which is needed for scheduling.

This adds a display name to the invoker's instance id, so that the operator of such a system can directly correlate the scheduling entity (topic name) and the corresponding container name in her system.

This display name differs from the "unique name" in that the latter might be something more cryptical like the node's host ip, to stay stable in between redeployments of the system. The unique name is not necessarily useful for the operator though.