Validate the Controller Swagger spec matches the impl (#3878)
This change wires up all existing API tests that use WskRestOperations
to validate the HTTP requests and responses against the Controller
API's Swagger spec.

It also adds a new Gradle task - :tests:testSwaggerCodegen - that does
a basic validation of the Swagger spec and generates a basic Java
client from that spec and verifies that the client compiles. CI is
wired up to run this new test task.

The result of both of the above are a number of changes to the Swagger
spec, also included here. This brings our Swagger spec much closer to
matching the reality of the Controller API, and any additional
coverage in the API test suite will further increase coverage of the
Swagger spec.

There are some existing tests that purposely test API responses when
passed invalid data, so the new SwaggerValidator trait includes a
small whilelist of invalid requests and responses to accept so the
build does not fail in these scenarios.

Based on review feedback, this eagerly converts all HTTP entities to
their Strict variants which allows both the SwaggerValidator and
downstream consumers to read the request/response bodies without
making a copy.

This also exposed another slight change needed in apiv1swagger.json
where rule enabling/disabling can sometimes return a text/plain body
instead of json.