Add additional debug information for intermittently failing ContainerProxyTests (#4278)
The `ContainerProxyTests` are failing intermittently in the pipeline. All of these intermittent test failures have either a wrong initialisation-time or it is missing completely.
Our theory currently is, that the Buffer in `acker` is called in the wrong order, as these calls are made asynchronously. If that theory is correct, this can be proven with this additional output.