Revert back to the previous version, as the new base image causes regression in performance tests. (#4353)
Reverting back to adoptopenjdk/openjdk8:x86_64-alpine-jdk8u172-b11 from adoptopenjdk/openjdk8:x86_64-alpine-jdk8u202-b08 (commit) as we see regression in Latency.

After briefly debugging, we found that the time was lost between Controller->Invoker & Invoker->Controller (as reverting version already helps to improve the problem there seems to be an issue with the communication with kafka). We didnt find a quick solution so reverting back to the old version, and in the meanwhile will continue to look for the root cause.